// src/models/ticket.ts
export class Ticket {
    id: string;
    refid: string;
    no: number;
    serialnumber: number;
    collection: number;
    award: number;
    game: string;
    date: string;
    time: string;
    isplayed: boolean;
    setnumber: number;
    num: number[];
    textAward: string;
    coinAward: number;
    ticketAward: number;
}